% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steves-package.R
\docType{package}
\name{steves-package}
\alias{steves}
\alias{steves-package}
\title{steves: Rick Steves' Europe Episodes for Teaching Data Analysis}
\description{
Tidy snapshot of every episode of the public-television travel series 'Rick Steves' Europe' (2000-2025), enriched with IMDB ratings, geocoded destinations, ISO country codes, episode thumbnails, and descriptive summaries. Designed as a companion dataset for introductory data analysis and visualization in the spirit of the 'moderndive' textbook: every row is an episode, every column is a candidate for a plot or a join.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://moderndive.github.io/steves/}
  \item \url{https://github.com/moderndive/steves}
  \item Report bugs at \url{https://github.com/moderndive/steves/issues}
}

}
\author{
\strong{Maintainer}: Chester Ismay \email{chester.ismay@gmail.com}

}
\keyword{internal}
